refactor: migrate inventory to feature module
