+++ src/features/inventory/state.js
@@ -1,0 +1,4 @@
+export const inventoryState = {
+  inventory: [],
+  equipment: { mainhand: { key: 'fist', type: 'weapon' }, head: null, torso: null, food: null },
+};
