+++ src/features/inventory/ui/weaponChip.js
@@ -1,5 +1,5 @@
 import { S } from '../../../game/state.js';
-import { WEAPON_FLAGS } from '../data/weapons.js';
+import { WEAPON_FLAGS } from '../../weaponGeneration/data/weapons.js';
 
 const weaponFeatureEnabled = Object.keys(WEAPON_FLAGS).some(w => w !== 'fist' && WEAPON_FLAGS[w]);
